requirements.txt
feedparser>=6.0.0
requests>=2.31.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
apscheduler>=3.10.4
